Remove Old Dev UI: Kubernetes
