--- extensions/kubernetes/vanilla/deployment/src/main/resources/dev-templates/embedded.html
@@ -1,3 +1,0 @@
-<a href="{urlbase}/manifests" class="badge badge-light">
-    <i class="fa fa-rocket fa-fw"></i>
-    Kubernetes Manifests</a>
