--- extensions/kubernetes/vanilla/deployment/src/main/resources/dev-templates/manifests.html
@@ -1,41 +1,0 @@
-{#include main fluid=true}
-{#title}Kubernetes Manifests{/title}
-{#styleref}
-<link rel="stylesheet" href="{devRootAppend}/resources/codemirror/lib/codemirror.css">
-{/styleref}
-{#scriptref}
-<script src="{devRootAppend}/resources/codemirror/lib/codemirror.js"></script>
-<script src="{devRootAppend}/resources/codemirror/mode/yaml/yaml.js"></script>
-<script src="{devRootAppend}/resources/codemirror/addon/selection/active-line.js"></script>
-{/scriptref}
-{#script}
-    var editor = CodeMirror.fromTextArea(document.getElementById("code"), {
-        mode: "yaml",
-        readOnly: true,
-        styleActiveLine: true,
-        lineNumbers: true,
-        lineWrapping: true,
-    });
-    editor.setSize("100%", "100%");
-{/script}
-{#body}
-
-{#for manifest in info:holder.manifests.orEmpty}
-
-
-<div class="row">
-    <div class="col-sm-12">
-        <div class="card">
-            <div class="card-header text-white font-weight-bold bg-primary">
-                {manifest.name}
-            </div>
-            <div class="card-body p-0">
-                <textarea id="code" name="code">{manifest.content}</textarea>
-            </div>
-        </div>
-    </div>
-</div>
-<br/>
-{/for}
-{/body}
-{/include}
